requirements.txt
torch
transformers
datasets
peft
accelerate
bitsandbytes
trl
safetensors
ipywidgets
huggingface_hub
python-dotenv
wandb
einops
scipy